Add new git command, refactor FFI code
